fix(#85): HiFi-GAN dtype — keep float32, cast mel.float() before hift
HiFi-GAN (stepaudio2 SineGenerator) generates float32 sine waves
internally. If the model is moved to BFloat16, l_linear weights become
BF16 but sine_wavs input stays float32 → RuntimeError dtype mismatch.

Fix: load HiFi-GAN with .float().to(device) and cast mel to float32
before calling hift().

+++ vllm_omni/model_executor/models/minicpm_o/minicpm_o_code2wav.py
@@ -117,7 +117,10 @@ class MiniCPMOCode2Wav(nn.Module):
             k.replace("generator.", ""): v for k, v in hift_ckpt.items()
         }
         hift.load_state_dict(hift_state_dict, strict=True)
-        hift.to(device).eval()
+        # HiFi-GAN uses float32 internally (SineGenerator produces float32 sine
+        # waves that must match linear layer dtypes).  Keep in float32 regardless
+        # of the main model dtype.
+        hift.float().to(device).eval()
         self.__dict__["_hift"] = hift
 
         logger.info(
@@ -176,7 +179,7 @@ class MiniCPMOCode2Wav(nn.Module):
                     embedding=embedding,
                     n_timesteps=n_timesteps,
                 )
-                wav = hift(mel)  # [1, 1, audio_len]
+                wav = hift(mel.float())  # HiFi-GAN requires float32
 
             results.append(wav)
